Corrección de la cadena de conexión. Cambio del tipo de dato de double a decimal. Se ha apreciado que si no asigno el resultado de miQuery a la lista de objetos, en caso de no haber resultados devuelve todo. Corregido. Filtros y llamadas OK

+++ Models/QueryParamsFilters.cs
@@ -39,8 +39,8 @@ public class QueryParamsFilters {
 
 
     //Filtros específicos para tipos de IVA
-    public double? filtroTasaMinima { get; set; }
-    public double? filtroTasaMaxima { get; set; }
+    public decimal? filtroTasaMinima { get; set; }
+    public decimal? filtroTasaMaxima { get; set; }
     public string? filtroDescripcionTipoIVA { get; set; }
     public DateTime? filtroFechaCreacionDesde { get; set; }
     public DateTime? filtroFechaCreacionHasta { get; set; }
